requirements.txt
feedparser==6.0.11
newspaper3k==0.2.8
transformers==4.41.0
accelerate==0.29.3
langchain==0.2.0
crewai==0.0.17
rake-nltk==1.0.6
python-frontmatter==1.0.0
requests==2.32.2
pillow==10.3.0
gitpython==3.1.43 